+++ src/Chinchilla/ISubscriptionState.cs
@@ -1,0 +1,10 @@
+namespace Chinchilla
+{
+    /// <summary>
+    /// A subscription state is a breakdown of what the current subscription is doing
+    /// </summary>
+    public class SubscriptionState
+    {
+
+    }
+}